% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbuds.R
\name{area}
\alias{area}
\title{Compute area of a polygon}
\usage{
area(X)
}
\arguments{
\item{X}{a polygon dataframe of x and y coordinates}
}
\value{
area of the polygon x.
}
\description{
This  function computes the area of a polygon X.
}
\examples{
filename<-system.file('data/YET629_02_w1L488nm-L561nm_sequence-10000.zip', package='yeast')
roi <- read.ijzip(filename)

cordi <- roi$`001_001`$coords
X <-  data.frame(X=cordi[,1], Y=cordi[,2])
area<-function(X){
 X<-cbind(X$X, X$Y)
 X<-rbind(X,X[1,])
 x<-X[,1]; y<-X[,2]; lx<-length(x)
 -sum((x[2:lx]-x[1:lx-1])*(y[2:lx]+y[1:lx-1]))/2
}

area(X)
}
