% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustercells.R
\name{cluster.cells}
\alias{cluster.cells}
\title{Cluster cells based on shape}
\usage{
cluster.cells(yeastCells, K = 3)
}
\arguments{
\item{x}{an ROI object from read.ijzip().}

\item{THRESHOLD}{an integer value specifying the area in pixels needed for an overlap to count as mother/budded relation. Default value is 100.}

\item{borderCol}{color value for overlap boundary.}
}
\value{

}
\description{
This function loads ImageJ roi zip files.
It returns a list object with rois normalized into coordinate system as well as mother/bud relations.
}
\examples{
filename<-system.file('data/YET629_02_w1L488nm-L561nm_sequence-10000.zip', package='yeast')
roi <- read.ijzip(filename)
yeastCells<-get.buds(roi)
cluster.cells(yeastCells)
}
