% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbuds.R
\name{getIntersection}
\alias{getIntersection}
\title{Find intersection of two objects}
\usage{
getIntersection(a, b)
}
\arguments{
\item{a}{first matrix object}

\item{b}{second matrix object}
}
\value{
the coordinates of intersection
}
\description{
This  function identifies the positions of intersection of two matrix objects.
}
