% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackCell.R
\name{trackCell}
\alias{trackCell}
\title{Imports ImageJ ROIs and control cell ID}
\usage{
trackCell(folder, ...)
}
\arguments{
\item{folder}{.zip folder containg .roi objects}

\item{full.names}{a logical value.  If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.Defaul TRUE.}
}
\value{
a list object of rois with corrected IDs adjusted to previous time stack, if needed.
}
\description{
This function loads ImageJ roi zip files.
It returns a list object with corrected cell IDs if needed.
}
\examples{
folder<-system.file('data/my.zipfiles', package='yeast')
yeastMovie<-trackCell(folder)
}
