% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbuds.R
\name{get.buds}
\alias{get.buds}
\title{Import and normalize ImageJ ROIs to cell coordinates}
\usage{
get.buds(
  x,
  THRESHOLD = 80,
  borderCol = rgb(0, 0, 0, 0.2),
  fillCol = NA,
  add = FALSE,
  xlab = "",
  ylab = "",
  main = "get buds",
  asp = 1,
  ...
)
}
\arguments{
\item{x}{.roi object from read.ijzip().}

\item{THRESHOLD}{an integer value specifying the area in pixels needed for an overlap to count as mother/budded relation. Default value is 100.}

\item{borderCol}{color value for overlap boundary.}

\item{add}{plot in the same coordinate system. default FALSE}

\item{xlab}{x axis label. Default empty.}

\item{ylab}{y axis label. Default empty.}

\item{main}{header. Default "get buds"}

\item{asp}{set the aspect ratio of the plot. Default 1.}

\item{fillcol}{color value for overlap filling}
}
\value{

}
\description{
This function loads ImageJ roi zip files.
It returns a list object with rois normalized into coordinate system as well as mother/bud relations.
}
\examples{
filename<-system.file('data/YET629_02_w1L488nm-L561nm_sequence-10000.zip', package='yeast')
roi <- read.ijzip(filename)
yeastCells<-get.buds(roi)
}
