% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbuds.R
\name{getPrincipalAxes}
\alias{getPrincipalAxes}
\title{getPrincipalAxes}
\usage{
getPrincipalAxes(contour, plot = F)
}
\arguments{
\item{contour}{an ROI object from read.ijzip().}

\item{plot}{a boolean value specifying whether to plot the ROI including PC1 and PC2.}
}
\value{
Returns one matrix for each principal component where the first row represents x,y coordinates for the end of the vector with lowest x value. The second row represents x,y coordinates for the end of the PC2 vector with highest x value. It also returns a matrix with the x,y coordinates for where PC1 and PC2 intersect, which represents the centroid of the polygon.
}
\description{
This function loads the contour of a cell and performs a principal component analysis.
It returns a list object with principal component 1 and principal component 2  as well as their intersection.
}
\examples{
filename<-system.file('data/YET629_02_w1L488nm-L561nm_sequence-10000.zip', package='yeast')
roi <- read.ijzip(filename)
PCAresult<-getPrincipalAxes(roi$`001_005`$coords, plot=T)
}
